refactor: modularize APA lesson domain module
