+++ packages/domain/src/apa-lesson/services/index.ts
@@ -1,0 +1,1 @@
+export * from "./validation";
